refactor: fold async calls into the sync files
