+++ python/mirascope/llm/calls/decorator.py
@@ -26,10 +26,8 @@ from ..tools import (
     ContextToolT,
     ToolT,
 )
-from .async_call import AsyncCall
-from .async_context_call import AsyncContextCall
-from .call import Call
-from .context_call import ContextCall
+from .call import AsyncCall, Call
+from .context_call import AsyncContextCall, ContextCall
 
 if TYPE_CHECKING:
     from ..clients import (
